userId 类型错误的问题

+++ common/userService/userService.go
@@ -38,7 +38,7 @@ func NewUserAndSave(unionId, openId, wxNickName, headUrl string, sex int32, city
 
 	//构造user
 	user := new(ddproto.User)
-	user.Id = proto.Uint32(id)
+	user.Id = proto.Uint32(uint32(id))
 	user.Sex = proto.Int32(sex)
 	user.City = proto.String(city)
 	user.Diamond = proto.Int64(NEW_USER_DIAMOND_REWARD)                //新用户注册的时候,默认的钻石数量
